spawn a base node in the cucumber test
