+++ applications/tari_validator_node/tests/features/json_rpc.feature
@@ -2,5 +2,6 @@ Feature: JSON-RPC methods
 
   @serial
   Scenario: The validator node returns its identity
-    Given a validator node "vn1"
+    Given a base node "bn1"
+    Given a validator node "vn1" connected to base node "bn1"
     Then the validator node "vn1" returns a valid identity
