+++ applications/tari_validator_node/tests/utils/mod.rs
@@ -1,0 +1,2 @@
+pub mod base_node;
+pub mod validator_node;
