+++ dan_layer/storage_lmdb/Cargo.toml
@@ -7,7 +7,8 @@ edition = "2021"
 
 [dependencies]
 tari_dan_engine = { path = "../engine" }
-tari_storage = {  git = "https://github.com/tari-project/tari.git", branch = "feature-dan", package = "tari_storage" }
+# tari_storage = {  git = "https://github.com/tari-project/tari.git", branch = "feature-dan", package = "tari_storage" }
+tari_storage = { path = "../../../tari/infrastructure/storage" }
 
 borsh = "0.9.3"
 lmdb-zero = "0.4.4"
